Finish up methods
