+++ src/ios/removeFromHeaderSearchPaths.js
@@ -1,6 +1,10 @@
+const mapHeaderSettings = require('./headerSearchPathIter');
+
 /**
  * Given Xcode project and absolute path, it makes sure there are no headers referring to it
  */
 module.exports = function addToHeaderSearchPaths(project, path) {
-  return null;
+  mapHeaderSettings(project, (headerSearchPaths) =>
+    headerSearchPaths.filter(searchPath !== path)
+  );
 };
