+++ src/ios/removeFromStaticLibraries.js
@@ -1,0 +1,19 @@
+const PbxFile = require('xcode/lib/pbxFile');
+
+/**
+ * Removes file from static libraries
+ *
+ * Similar to `node-xcode` addStaticLibrary
+ */
+module.exports = function removeFromStaticLibraries(project, path, opts) {
+  const file = new PbxFile(path);
+
+  file.target = opts ? opts.target : undefined;
+
+  project.removeFromPbxFileReferenceSection(file);
+  project.removeFromPbxBuildFileSection(file);
+  project.removeFromPbxFrameworksBuildPhase(file);
+  project.removeFromLibrarySearchPaths(file);
+
+  return file;
+};
